+++ browser/modules/draw.js
@@ -148,7 +148,7 @@ module.exports = {
             return;
         }
         // Switch on Conflict
-        // $('#main-tabs a[href="#conflict-content"]').trigger('click');
+        $('#main-tabs a[href="#conflict-content"]').trigger('click');
         conflictSearch.makeSearch("Fra tegning", null, selectedDrawing, true);
     },
 
@@ -156,7 +156,7 @@ module.exports = {
 
     makeConflictSearchWithAll: () => {
         // Switch on Conflict
-        // $('#main-tabs a[href="#conflict-content"]').trigger('click');
+        $('#main-tabs a[href="#conflict-content"]').trigger('click');
         conflictSearch.makeSearch("Fra tegning", null, null, true);
     },