+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [CalVer](https://calver.org/).
 - Conflict search module will now search protected layers if user is signed in.
 - On `zoomToExtentOfgeoJsonStore` in geocloud.js pan map one pixel to defeat a strange bug, which causes a browser freeze.
 - Layer filter auto complete menu is now postioned absolute instead of fixed, so it will not hang in the same place when scrolling layer tree.
+- Composit QGIS Layers now work. They are not longer bypassed by MapServer where the merge request is done.
 
 ## [2021.5.0] - 2021-4-5
 ### Changed